integrations/ollama/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ollama-deepstack"
dynamic = ["version"]
description = 'An integration between the Ollama LLM framework and Deepstack'
readme = "README.md"
requires-python = ">=3.8"
license = "Apache-2.0"
keywords = []
authors = [
    { name = "Alistair Rogers", email = "alistairlr112@gmail.com" },
    { name = "Sachin Sachdeva", email = "emailforsachinsachdeva@gmail.com" },
    { name = "khulnasoft GmbH", email = "info@khulnasoft.com" },
]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["deepstack-ai", "requests"]

[project.urls]
Documentation = "https://github.com/khulnasoft/deepstack-core-integrations/tree/main/integrations/ollama#readme"
Issues = "https://github.com/khulnasoft/deepstack-core-integrations/issues"
Source = "https://github.com/khulnasoft/deepstack-core-integrations/tree/main/integrations/ollama"

[tool.hatch.build.targets.wheel]
packages = ["src/deepstack_integrations"]

[tool.hatch.version]
source = "vcs"
tag-pattern = 'integrations\/ollama-v(?P<version>.*)'

[tool.hatch.version.raw-options]
root = "../.."
git_describe_command = 'git describe --tags --match="integrations/ollama-v[0-9]*"'

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
    "deepstack-pydoc-tools",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]
cov = [
    "test-cov",
    "cov-report",
]
docs = [
  "pydoc-markdown pydoc/config.yml"
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]


[tool.hatch.envs.lint]
detached = true
dependencies = [
    "black>=23.1.0",
    "mypy>=1.0.0",
    "ruff>=0.0.243",
]

[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive --explicit-package-bases {args:src/ tests}"
style = [
    "ruff {args:.}",
    "black --check --diff {args:.}",
]
fmt = [
    "black {args:.}",
    "ruff --fix {args:.}",
    "style",
]
all = [
    "style",
    "typing",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff.isort]
known-first-party = ["src"]

[tool.black]
target-version = ["py38"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py38"
line-length = 120
select = [
    "A",
    "ARG",
    "B",
    "C",
    "DTZ",
    "E",
    "EM",
    "F",
    "I",
    "ICN",
    "ISC",
    "N",
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    "Q",
    "RUF",
    "S",
    "T",
    "TID",
    "UP",
    "W",
    "YTT",
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Ignore checks for possible passwords
    "S105", "S106", "S107",
    # Ignore complexity
    "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
    # Don't touch unused imports
    "F401",
]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "parents"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
# Examples can print their output
"examples/**" = ["T201"]


[tool.coverage.run]
source = ["deepstack_integrations"]
branch = true
parallel = false

[tool.coverage.report]
omit = ["*/tests/*", "*/__init__.py"]
show_missing=true
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as slow (deselect with '-m \"not integration\"')",
]
addopts = [
    "--import-mode=importlib",
]

[[tool.mypy.overrides]]
module = [
    "deepstack.*",
    "deepstack_integrations.*",
    "pytest.*"
]
ignore_missing_imports = true
